backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community
langchain-core
langchain-ollama
langchain-chroma
chromadb
pydantic